Added signature check functionality
included checkSignature Function which is used in the Authentication Request. Input the known address of the Authenticator and  msgHash which is decrypted using the eth built in function ecrecover. Compare the two and if matched and output is True then continue with the operation. Note inputs also required are v, r, s which are the mathematical variables required for the decryption and provided with the msgHash.